requirements.txt
streamlit
ollama
python-docx
fpdf
pyttsx3
speechrecognition
pyaudio
